Make healer & wizard regen bonus more gradual
Yes, regen is too slow in vanilla. But the force-bolt-slinging
that a level 1 wizard is capable of in FIQhack is a little crazy.
Let the bonus grow more gradually, and let it grow to a higher
maximum.

I'm not sure whether healers "needed" more regen, but it seemed
cleaner to change them both at the same time in the same way. As
a side benefit, gaining levels will have a more noticable effect!