Further racialize starting inventory
Humans get a "canteen" of alignment-appropriate water.

Elves ALWAYS get an instrument, and it's always tonal!

Gnomes get an assortment of useful objects, including a leash,
which is no longer a tourist thing. It's a gnome thing.

Orcs get a lil' bit of meat and poison. Do not combine these!

Dwarves get booze. Obviously.