rejigger wand skill caps based on my own opinion of where they should be :)
This is partly just a matter of taste, but also a matter of
reducing what feels like a "skill tax" on some roles that already
have a lot of places to invest their skill points.

I would bump wizard down one notch, but people would be angry and
wands + wizards is inarguable flavor wise.