[FOURK] [UNFINISHED] Import the magic chests feature from DynaHack.
TODO: figure out wtf to do with restore.c

Magic chests: dungeon feature with shared contents

Any item put into a magic chest can be looted out of any other magic
chest in the dungeon, placed in set locations.  This eliminates
tedious stash consolidation and transporting of items to and from
stashes and fixed dungeon resources like altars, shops and water.

The back-end design of the shared magic chest inventory is much like
how migrating objects used to be handled before they were removed: an
object chain whose contents are considered global as opposed to local
(level-based), meaning timers in a magic chest follow the player
across levels instead of staying on any one level.

Most of the code here is dedicated to hacking in magic chests into the
container looting system under the guise of `&zeroobj`, since magic
chests are terrain and not objects.

The map character for magic chests is an open parenthesis.  I would
have preferred `C` or `M` for the `*.des` files, but both were already
taken (by clouds and muddy swamps respectively).

Wishing for "magic chest" in wizard mode creates a magic chest at your
location, and the `#stats` command (again in wizard mode) shows the
count of objects and memory usage of the magic_chest_objs chain.

There are some differences in Fourk, compared to DynaHack.  The most
important of these is that some of the Minetown levels (the ones
without maps) currently don't have a magic chest in Fourk, because the
feature DynaHack used to place them there (TERRAIN) does not appear to
be supported by our levcomp.  (That may be a feature of paxed's
levcomp rewrite, which is definitely in UnNetHack and so may well be
in Dyna as well, which borrows a lot of things from Un.)  Also, some
levels have the chest in a slightly different location in Fourk.
Finally, there are some implementation details that differ due to
divergent codebases.