make cornuthaums work for all gnomes OR wizards
this was someone else's YANI, but I'm not sure whose. I like it.

this commit also bugfixes one of fiddlehack's earliest commits,
which was supposed to make cornuthaums add +1 int and wis for
wizards, but due to misuse of the "if" clause, it made cornuthaums
simply not do anything to attributes regardless of the wearer.