halve the to-hit from luck and XP level
towards a goal of making dex, status effects, kebabable, weapon
skill, and all manner of other to-hit bonuses actually matter, the
huge flat to-hit sources of luck and XP will be halved.

A small +1 is added to half your level so that the start of the
game, XP 1, isn't changed.